+++ src/components/Dropdown/Dropdown.tsx
@@ -20,8 +20,8 @@ import React, { ReactElement, ReactNode, useCallback, useMemo, useState } from '
 import { Dropdown as AntdDropdown } from 'antd'
 import classNames from 'classnames'
 
-import { AntdMenuClickEvent, AntdMenuItem, AntdMenuItems, Placement, antdSourceMap } from './Dropdown.types'
-import { DropdownClickEvent, DropdownItem, DropdownProps, DropdownTrigger, ItemLayout } from './Dropdown.props'
+import { AntdMenuClickEvent, AntdMenuItem, AntdMenuItems, Placement, antdSourceMap } from './types'
+import { DropdownClickEvent, DropdownItem, DropdownProps, DropdownTrigger, ItemLayout } from './props'
 import { Footer, useFooterWithHookedActions } from './components/Footer'
 import { Divider } from '../Divider'
 import Label from './components/Label'
